% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransactionHandler.R
\name{TransactionHandler}
\alias{TransactionHandler}
\title{Transaction Handler}
\usage{
TransactionHandler(
  inputId,
  label,
  contract_address,
  contract_abi,
  contract_method,
  args,
  enabled,
  default = ""
)
}
\description{
Given the configuration arguments, this function will prepare an EVM transaction
and display a button that will run the transaction when clicked.
}
