% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WalletHandler.R
\name{WalletHandler}
\alias{WalletHandler}
\title{Wallet Handler}
\usage{
WalletHandler(inputId, chainId, default = "")
}
\description{
Manages the wallet connection and network switching for EVM wallets.
}
